+++ src/IdGenerator.ts
@@ -1,0 +1,9 @@
+import uuid from 'react-native-uuid';
+
+class ReactNativeIdGenerator {
+  newId(): string {
+    return uuid.v4();
+  }
+}
+
+export { ReactNativeIdGenerator };
