+++ src/index.tsx
@@ -1,8 +1,8 @@
 import { defineBKTConfig } from 'bkt-js-client-sdk';
 import type { BKTConfig, RawBKTConfig } from 'bkt-js-client-sdk';
-import uuid from 'react-native-uuid';
 import { SDK_VERSION } from './version';
 import { SOURCE_ID_REACT_NATIVE } from './SourceId';
+import { ReactNativeIdGenerator } from './IdGenerator';
 
 export * from 'bkt-js-client-sdk';
 export * from 'bkt-react-client-sdk';
@@ -16,10 +16,4 @@ export function defineReactNativeBKTConfig(config: RawBKTConfig): BKTConfig {
     wrapperSdkVersion: SDK_VERSION,
     userAgent: `Bucketeer React Native SDK(${SDK_VERSION})`,
   });
-}
-
-class ReactNativeIdGenerator {
-  newId(): string {
-    return uuid.v4();
-  }
 }
